+++ inst/resources/init-rprofile.R
@@ -1,3 +1,3 @@
-#### -- Packrat Autoloader (version 0.4.9-9) -- ####
+#### -- Packrat Autoloader (version 0.4.9-11) -- ####
 source("packrat/init.R")
 #### -- End Packrat Autoloader -- ####